map/requirements.txt
Django~=4.1.7
environs~=9.5.0
django-admin-sortable2~=2.1.4
Pillow~=9.4.0
django-tinymce~=3.5.0
dj-database-url~=1.2.0
requests~=2.28.2
